mrc-attention-backend-setup(extract-attention-backend-setup-move,mechanical_provable): move init_attention_backends chain to attention_backend_setup module (cut+paste)
PR-Title: Extract attention-backend setup into a module
