+++ python/sglang/srt/model_executor/model_runner.py
@@ -110,6 +110,11 @@ from sglang.srt.model_executor.forward_context import (
 from sglang.srt.model_executor.graph_shared_output import GraphSharedOutput
 from sglang.srt.model_executor.hook_manager import register_forward_hooks
 from sglang.srt.model_executor.model_runner_components import misc_utils
+from sglang.srt.model_executor.model_runner_components.attention_backend_setup import (
+    build_attention_backends,
+    configure_aux_hidden_state_capture,
+    get_attention_backend,
+)
 from sglang.srt.model_executor.model_runner_components.kv_pool_runtime import (
     compute_post_capture_kv_resize,
     is_post_capture_kv_active,
@@ -222,20 +227,6 @@ elif current_platform.is_out_of_tree():
 
 
 logger = logging.getLogger(__name__)
-
-
-class ResolvedAttentionBackendStr(msgspec.Struct, frozen=True, kw_only=True):
-    prefill: str
-    decode: str
-    is_draft_override: bool = False
-
-
-class AttentionBackends(msgspec.Struct, frozen=True, kw_only=True):
-    attn_backend: AttentionBackend
-    decode_attn_backend: Optional[AttentionBackend]
-    decode_attn_backend_group: list[AttentionBackend]
-    prefill_attention_backend_str: str
-    decode_attention_backend_str: str
 
 
 @dataclass
@@ -738,7 +729,7 @@ class ModelRunner:
         """Initialize attention backends only (no cuda graph capture)."""
         # Must be called BEFORE init_decode_cuda_graph() so CUDA graph capture
         # runs with aux hidden state capture enabled.
-        ModelRunner.configure_aux_hidden_state_capture(
+        configure_aux_hidden_state_capture(
             model=self.model,
             eagle_use_aux_hidden_state=self.spec_aux_config.eagle_use_aux_hidden_state,
             eagle_aux_hidden_state_layer_ids=self.spec_aux_config.eagle_aux_hidden_state_layer_ids,
@@ -746,203 +737,12 @@ class ModelRunner:
             dflash_target_layer_ids=self.spec_aux_config.dflash_target_layer_ids,
             is_dspark=self.spec_algorithm.is_dspark(),
         )
-        backends = ModelRunner.build_attention_backends(model_runner=self)
+        backends = build_attention_backends(model_runner=self)
         self.attn_backend = backends.attn_backend
         self.decode_attn_backend = backends.decode_attn_backend
         self.decode_attn_backend_group = backends.decode_attn_backend_group
         self.prefill_attention_backend_str = backends.prefill_attention_backend_str
         self.decode_attention_backend_str = backends.decode_attention_backend_str
-
-    @staticmethod
-    def configure_aux_hidden_state_capture(
-        *,
-        model,
-        eagle_use_aux_hidden_state: bool,
-        eagle_aux_hidden_state_layer_ids,
-        dflash_use_aux_hidden_state: bool,
-        dflash_target_layer_ids,
-        is_dspark: bool,
-    ) -> None:
-        """Configure auxiliary hidden state capture for speculative decoding.
-
-        Must be called before CUDA graph capture so the captured graphs
-        include aux hidden state output paths.
-        """
-        if eagle_use_aux_hidden_state:
-            model.set_eagle3_layers_to_capture(eagle_aux_hidden_state_layer_ids)
-        if dflash_use_aux_hidden_state:
-            if is_dspark and hasattr(model, "set_dspark_layers_to_capture"):
-                model.set_dspark_layers_to_capture(dflash_target_layer_ids)
-            elif hasattr(model, "set_dflash_layers_to_capture"):
-                model.set_dflash_layers_to_capture(dflash_target_layer_ids)
-            else:
-                raise ValueError(
-                    f"Model {model.__class__.__name__} implements neither "
-                    "set_dspark_layers_to_capture nor set_dflash_layers_to_capture, "
-                    "one of which is required for DFLASH/DSPARK."
-                )
-
-    @staticmethod
-    def build_attention_backends(*, model_runner: ModelRunner) -> AttentionBackends:
-        """Init attention kernel backend."""
-        server_args = model_runner.server_args
-
-        # TODO: Refactor device-specific init branches into platform interface (separate PR).
-        if model_runner.device in ("cuda", "musa"):
-            init_cublas()
-
-        resolved = _resolve_attention_backend_strs(
-            server_args=server_args, is_draft_worker=model_runner.is_draft_worker
-        )
-
-        if server_args.enable_pdmux:
-            attn_backend = _build_resolved_backend(
-                model_runner=model_runner, resolved=resolved, init_new_workspace=True
-            )
-            decode_attn_backend_group = [
-                _build_resolved_backend(
-                    model_runner=model_runner,
-                    resolved=resolved,
-                    init_new_workspace=False,
-                )
-                for _ in range(server_args.sm_group_num)
-            ]
-            decode_attn_backend = decode_attn_backend_group[0]
-        elif server_args.enable_two_batch_overlap and not model_runner.is_draft_worker:
-            attn_backend = TboAttnBackend.init_new(
-                lambda: _build_resolved_backend(
-                    model_runner=model_runner,
-                    resolved=resolved,
-                    init_new_workspace=False,
-                )
-            )
-            decode_attn_backend = None
-            decode_attn_backend_group = []
-        else:
-            attn_backend = _build_resolved_backend(
-                model_runner=model_runner, resolved=resolved, init_new_workspace=False
-            )
-            decode_attn_backend = None
-            decode_attn_backend_group = []
-
-        if (
-            model_runner.device == "npu"
-            and envs.SGLANG_ZBAL_LOCAL_MEM_SIZE.get() > 0
-            and not model_runner.is_draft_worker
-        ):
-            # lazy init for zbal with mix mode (before graph capture when enable_cuda_graph)
-            from sglang.srt.hardware_backend.npu.utils import lazy_init_zbal_gva_mem
-
-            lazy_init_zbal_gva_mem(
-                model_runner.device,
-                model_runner.gpu_id,
-                get_world_group().rank_in_group,
-                get_world_group().world_size,
-                get_world_group().cpu_group,
-            )
-
-        # Record resolved per-mode backends on the backend for model dispatch.
-        attn_backend.prefill_attention_backend_str = resolved.prefill
-        attn_backend.decode_attention_backend_str = resolved.decode
-
-        return AttentionBackends(
-            attn_backend=attn_backend,
-            decode_attn_backend=decode_attn_backend,
-            decode_attn_backend_group=decode_attn_backend_group,
-            prefill_attention_backend_str=resolved.prefill,
-            decode_attention_backend_str=resolved.decode,
-        )
-
-    @staticmethod
-    def get_attention_backend(
-        *, model_runner: ModelRunner, init_new_workspace: bool = False
-    ) -> AttentionBackend:
-        """Init attention kernel backend."""
-        resolved = _resolve_attention_backend_strs(
-            server_args=model_runner.server_args,
-            is_draft_worker=model_runner.is_draft_worker,
-        )
-        return _build_resolved_backend(
-            model_runner=model_runner,
-            resolved=resolved,
-            init_new_workspace=init_new_workspace,
-        )
-
-    @staticmethod
-    def _resolve_attention_backend_strs(
-        *, server_args: ServerArgs, is_draft_worker: bool
-    ) -> ResolvedAttentionBackendStr:
-        draft_attn_backend = server_args.speculative_draft_attention_backend
-        if is_draft_worker and draft_attn_backend:
-            logger.warning(
-                f"Overriding draft attention backend to {draft_attn_backend}."
-            )
-            # Single backend for all draft modes (no prefill/decode split).
-            return ResolvedAttentionBackendStr(
-                prefill=draft_attn_backend,
-                decode=draft_attn_backend,
-                is_draft_override=True,
-            )
-        prefill, decode = server_args.get_attention_backends()
-        return ResolvedAttentionBackendStr(prefill=prefill, decode=decode)
-
-    @staticmethod
-    def _build_resolved_backend(
-        *,
-        model_runner: ModelRunner,
-        resolved: ResolvedAttentionBackendStr,
-        init_new_workspace: bool,
-    ) -> AttentionBackend:
-        if resolved.is_draft_override:
-            attn_backend = _build_backend_from_str(
-                model_runner=model_runner,
-                backend_str=resolved.prefill,
-                init_new_workspace=init_new_workspace,
-            )
-        elif resolved.decode != resolved.prefill:
-            from sglang.srt.layers.attention.hybrid_attn_backend import (
-                HybridAttnBackend,
-            )
-
-            attn_backend = HybridAttnBackend(
-                model_runner,
-                decode_backend=_build_backend_from_str(
-                    model_runner=model_runner,
-                    backend_str=resolved.decode,
-                    init_new_workspace=init_new_workspace,
-                ),
-                prefill_backend=_build_backend_from_str(
-                    model_runner=model_runner,
-                    backend_str=resolved.prefill,
-                    init_new_workspace=init_new_workspace,
-                ),
-            )
-            logger.info(
-                f"Using hybrid attention backend for decode and prefill: "
-                f"decode_backend={resolved.decode}, "
-                f"prefill_backend={resolved.prefill}."
-            )
-            logger.warning(
-                "Warning: Attention backend specified by --attention-backend or default backend might be overridden."
-                "The feature of hybrid attention backend is experimental and unstable. Please raise an issue if you encounter any problem."
-            )
-        else:
-            attn_backend = _build_backend_from_str(
-                model_runner=model_runner,
-                backend_str=model_runner.server_args.attention_backend,
-                init_new_workspace=init_new_workspace,
-            )
-        return attn_backend
-
-    @staticmethod
-    def _build_backend_from_str(
-        *, model_runner: ModelRunner, backend_str: str, init_new_workspace: bool
-    ) -> AttentionBackend:
-        if backend_str not in ATTENTION_BACKENDS:
-            raise ValueError(f"Invalid attention backend: {backend_str}")
-        model_runner.init_new_workspace = init_new_workspace
-        full_attention_backend = ATTENTION_BACKENDS[backend_str](model_runner)
-        return attn_backend_wrapper(model_runner, full_attention_backend)
 
     def init_cuda_graphs(self, capture_decode_cuda_graph: bool = True):
         """Capture cuda graphs. Requires init_attention_backends() to have run.
@@ -1288,7 +1088,7 @@ class ModelRunner:
             self._record_kv_cache_dtype(resolved_kv_cache_dtype)
 
     def _get_attention_backend(self, init_new_workspace: bool = False):
-        return ModelRunner.get_attention_backend(
+        return get_attention_backend(
             model_runner=self, init_new_workspace=init_new_workspace
         )
